refactor: split control plane resource handlers
